# opt: reduce allocations in optbuilder